requirements.txt
#Fast API
fastapi ==0.112.1
uvicorn ==0.30.6
pytest==5.3.2
requests==2.32.3

#Alembic migrations
alembic==1.13.1
asyncpg==0.29.0
socket.io-emitter==0.1.5.1


#Celery
#celery==5.2.3
#redis==3.5.3
#celery-once==3.0.1
#greenlet==1.1.2
#eventlet==0.26.1
#gevent==21.12.0

#env
python-dotenv==1.0.1


psycopg2-binary==2.9.9


#jwt
python-jose==3.3.0

SQLAlchemy==2.0.32
pydantic==1.10.17
passlib==1.7.4

jose==1.0.0